services/skills-kit/lyzr-kit/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lyzr-kit"
version = "0.3.0"
description = "Python SDK for managing AI agents, tools, and features via the Lyzr platform"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Lyzr", email = "support@lyzr.ai" }
]
keywords = ["ai", "agents", "llm", "lyzr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydantic>=2.0",
    "typer>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "websockets>=12.0",
    "prompt_toolkit>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.8",
    "mypy>=1.0",
]

[project.scripts]
lk = "lyzr_kit.main:app"

[project.urls]
Homepage = "https://github.com/LyzrCore/lyzr-kit"
Documentation = "https://github.com/LyzrCore/lyzr-kit"
Repository = "https://github.com/LyzrCore/lyzr-kit"

[tool.hatch.build.targets.wheel]
packages = ["src/lyzr_kit"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # Pyflakes
    "I",    # isort
    "W",    # pycodestyle warnings
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # Line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.coverage.run]
source = ["src/lyzr_kit"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]
